Move test, add cache check
